Remove use of Util.get*Instance()
Use the "Telehash" object as context for accessing the
crypto/storage/network implementations, and remove the
Util.get*Instance() methods.  In places where we dont have a reference
to a valid Telehash object, we use a ThreadLocal scheme to obtain one.
This isn't much better than the previous method, but it may be a step
closer to a real solution.  (I'd like to avoid use of a heavyweight
dependency injection system.)